Copy pipe function to strict module with different typing to create a separate strict version
